Refactor TLVTree to not have a View attribute
Reasoning: The mental model was a bit confusing, since
the Model\View reads the YAML and has a Tree attribute, while
this Tree had the same View as attribute.

But we only need some attributes from the View in the Tree,
might even be possible to use Model\View instead of Tree in some cases
and then use view->getTree().